French GDP

Dataset containing the quarterly growth of the total gross domestic product (GDP) of France and quarterly series of the French business climate in level and in difference.

Usage

gdp

Format

A quarterly ts object from 1949Q2 to 2024Q1.

Source

INSEE

Details

Dataset containing the quarterly growth of the total gross domestic product (GDP, "growth_gdp") of France, in volumes chained at previous year prices, seasonally and working day adjusted; and the French business climate in level.

The French business climate is a monthly series, it is transformed into three quarterly series using the month's place in the quarter. For example, "bc_fr_m1" contains the values in the first month of each quarter, and the "diff_fr_m1" is the difference of the previous variable (the 2000Q1 value corresponds to the difference in business climate between January 2000 and October 1999).

Data were downloaded March 15, 2024 and might therefore differ from the latest available data.