Residual Effect Regression

Checks there is no residual effect in a model

Usage

lm_residual_effect(x, var = c(-1))

Arguments

x

lm object

var

the variables on which the residuals are to be regressed. By default use them all and cancel the explained variable

Examples

data("AirPassengers")
model <- lm(AirPassengers ~ time(AirPassengers))
lm_residual_effect(model)
#> 
#> Call:
#> lm(formula = resid(x) ~ ., data = as.data.frame(x$model[, var]))
#> 
#> Coefficients:
#>      (Intercept)  `x$model[, var]`  
#>        8.013e-13        -4.096e-16  
#>