Extract data frame for lm_fixed_coeff

According to parameter fixed_var, computes a new explained variable, which is the explained variable minus the product between estimated coefficients and values of the fixed variables.

Usage

resid_lm_fixed(x, fixed_var)

Arguments

x

lm model

fixed_var

list of variables that don't vary through time according to hansen_test

Value

A new environment where the explained variable is named "fixed".