Business Surveys

Dataset containing the quarterly growth of production in the manufacturing sector and its main sub-sectors, quarterly balance of opinion of business surveys published by INSEE and Banque de France and quarterly overhang of the industrial production index.

Usage

manufacturing

Format

A quarterly ts object from 1949Q2 to 2024Q1.

Source

INSEE, Banque de France

Details

Dataset containing the quarterly growth of production in the manufacturing sector and its main sub-sectors and quarterly series series of business surveys published by INSEE and Banque de France.

The sectors studied are:

Manufacturing industry

Food products and beverages (C1)

Capital goods (C3)

Transport equipments (C4)

Other manufacturing (C5)

"manuf_prod" contains the quarterly growth in production in the manufacturing sector, and the sub-sectors are in the form "prod_c1", "prod_c3", "prod_c4" and "prod_c5".

The overhang of the industrial production index corresponds to the quarterly growth obtained extending the series by the last known value:

"overhang_ipi0" is the quarterly growth obtained extending the series by the last value of the previous quarter (December, March, June, September);

"overhang_ipi1" is the quarterly growth obtained extending the series by the first value of the current quarter (January, April, July, October);

"overhang_ipi2" is the quarterly growth obtained extending the series by the second value of the current quarter (February, May, August, November).

The business surveys being monthly, the balance of opinion are transformed into three quarterly series using the month's place in the quarter (for example taking the values of January, April, July and October). Variable names are constructed as the combination of several codes defined as follows:

data source code (INSEE, ins, or Banque de France, bdf);

name of the balance of opinion:

Code	Definition

bc	Business climate

oscd	Overall order books

tppa and prodpas	Past production

tppre and prodpre	Personal production expectation

sitcar	Situation of order books

evocar	Evolution of order books

prix	Selling prices

stocks	Inventories of finished goods

tres	Cash position

tuc	Capacity utilisation rate

sector: nothing for the manufacturing industri and "c1", "c3", "c4" or "c5" for the sub-sectors;

place of the month in the quarter: m1, m2 or m3 for the first, second or third month of the quarter.

The dataset also contains some dummies labelled "indYYYYQX", where YYYY is the year and X is the quarter.